docs: clarify MaterialKey mapping constraints
