+++ src/table_builder.rs
@@ -1,6 +1,6 @@
 use crate::material_key::MaterialKey;
 use crate::score::DtzScoreRange;
-use shakmaty::{CastlingMode, Chess, FromSetup, Piece, Position, Setup, Square};
+use shakmaty::Position;
 
 pub struct TableBuilder {
     material: MaterialKey,
@@ -9,7 +9,7 @@ pub struct TableBuilder {
 
 impl TableBuilder {
     pub fn new(material: MaterialKey) -> Self {
-        let positions = Self::total_positions(&material);
+        let positions = material.total_positions();
 
         Self {
             material,
@@ -29,7 +29,7 @@ impl TableBuilder {
     /// This performs one bellman update on every position in the table.
     fn step(&mut self) {
         for pos_index in 0..self.positions.len() {
-            let position = self.index_to_position(pos_index);
+            let position = self.material.index_to_position(pos_index);
 
             // If the position is invalid, skip it.
             if let Some(position) = position {
@@ -49,7 +49,7 @@ impl TableBuilder {
                     .map(|chess_move| {
                         let mut child_position = position.clone();
                         child_position.play_unchecked(chess_move);
-                        let child_index = self.position_to_index(&child_position);
+                        let child_index = self.material.position_to_index(&child_position);
                         self.positions[child_index]
                     })
                     .fold(self.positions[pos_index], |a, b| a.negamax(&b));
@@ -58,52 +58,12 @@ impl TableBuilder {
             }
         }
     }
-
-    fn total_positions(material: &MaterialKey) -> usize {
-        64usize.pow(material.len() as u32)
-    }
-
-    /// Make a position from a position index.
-    ///
-    /// Returns `None` if the position is invalid.
-    ///
-    /// TODO:
-    /// Don't assign indices to invalid positions.
-    /// Add support for duplicated material (e.g. 2 knights)
-    fn index_to_position(&self, mut pos_index: usize) -> Option<Chess> {
-        let mut setup = Setup::empty();
-
-        for piece in self.material.iter() {
-            let index = pos_index % 64;
-            let square = Square::new(index as u32);
-
-            if setup.board.piece_at(square).is_some() {
-                return None;
-            }
-
-            setup.board.set_piece_at(square, *piece);
-            pos_index /= 64;
-        }
-
-        Chess::from_setup(setup, CastlingMode::Standard).ok()
-    }
-
-    fn position_to_index(&self, position: &Chess) -> usize {
-        let mut index = 0;
-
-        for piece in self.material.iter().rev() {
-            let square = position.board().by_piece(*piece).first().unwrap();
-            index = index * 64 + square.to_usize();
-        }
-
-        index
-    }
 }
 
 #[cfg(test)]
 mod tests {
     use super::*;
-    use shakmaty::fen::Fen;
+    use shakmaty::{CastlingMode, Piece, fen::Fen};
 
     #[test]
     fn position_index_roundtrip() {
@@ -122,9 +82,12 @@ mod tests {
             .into_position(CastlingMode::Standard)
             .unwrap();
 
-        let index = tb.position_to_index(&position);
-        let reconstructed = tb.index_to_position(index).expect("valid position");
+        let index = tb.material.position_to_index(&position);
+        let reconstructed = tb
+            .material
+            .index_to_position(index)
+            .expect("valid position");
 
-        assert_eq!(tb.position_to_index(&reconstructed), index);
+        assert_eq!(tb.material.position_to_index(&reconstructed), index);
     }
 }
